A huge step towards 'streaming' - h264 Annex B stream is read from an InputStream and written as fMP4

+++ streaming/src/main/java/com/mp4parser/streaming/extensions/SampleFlagsSampleExtension.java
@@ -50,16 +50,16 @@ public class SampleFlagsSampleExtension implements SampleExtension {
         return sampleDependsOn;
     }
 
-    public void setSampleDependsOn(byte sampleDependsOn) {
-        this.sampleDependsOn = sampleDependsOn;
+    public void setSampleDependsOn(int sampleDependsOn) {
+        this.sampleDependsOn = (byte) sampleDependsOn;
     }
 
     public byte getSampleIsDependedOn() {
         return sampleIsDependedOn;
     }
 
-    public void setSampleIsDependedOn(byte sampleIsDependedOn) {
-        this.sampleIsDependedOn = sampleIsDependedOn;
+    public void setSampleIsDependedOn(int sampleIsDependedOn) {
+        this.sampleIsDependedOn = (byte) sampleIsDependedOn;
     }
 
     public byte getSampleHasRedundancy() {
